backend/requirements.txt
fastapi
uvicorn
python-socketio
python-dotenv
langchain
langchain-openai
langgraph
pydantic
groq
